backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.0.1
sympy==1.12
numpy==1.24.3
google-generativeai==0.3.1
requests==2.31.0
wolframalpha==5.0.0
regex==2023.8.8
opencv-python-headless==4.8.0.76